# [CRuby] Optimize `strscan_do_scan()`: Remove unnecessary use of `rb_enc_get()`